tds_ssl::ssl_read_cb: fix clang warning

+++ src/ssl.cpp
@@ -314,7 +314,7 @@ namespace tds {
 
             rb.read(span(sp.data(), to_copy));
 
-            return to_copy;
+            return (int)to_copy;
         } else {
             int copied = 0;
